+++ com.peopleground.sagwim.fe/src/pages/admin/AdminChartsPage.module.css
@@ -1,0 +1,19 @@
+/* ── AdminChartsPage ── */
+
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: var(--sp-6);
+}
+
+.chartsGrid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  gap: var(--sp-5);
+}
+
+@media (max-width: 1024px) {
+  .chartsGrid {
+    grid-template-columns: 1fr;
+  }
+}
